Refactoring the server code
