+++ type4py/server/response.py
@@ -1,0 +1,7 @@
+class ServerResponse:
+    def __init__(self, response: dict, error: str=None) -> None:
+        self.response = response
+        self.error = error
+    
+    def get(self):
+        return {'response': self.response, 'error': self.error}
