+++ type4py/server/supervisord.conf
@@ -173,7 +173,7 @@ serverurl=unix:///tmp/supervisor.sock ; use a unix:// URL  for a unix socket
 
 [program:type4py_server]
 directory=/home/amir/projects/type4py/type4py/server
-command=gunicorn --reload -b localhost:5000 --log-level=DEBUG -w 4 -k gevent wsgi:app --timeout 120
+command=gunicorn --reload -b localhost:5000 --log-level=DEBUG -w 1 -k gevent "wsgi:app" --timeout 120
 environment=PATH="/home/amir/py_venv/py38/bin/:%(ENV_PATH)s"
 autostart=true
 autorestart=true
