+++ type4py/server/wsgi.py
@@ -1,4 +1,3 @@
-from logging import debug
 from type4py.server.app import app
 
 if __name__ == "__main__":
